csrrd csrwr works
add cpu7_csr module

it only has crme.ie

Ther is a bug that when csrrd follows right after csrwr, it's like
the situation where the data forwarding is needed for csr register.

Read-After-Write  RAW

In this commit, the byp logic for csr is added.

But, sutmitting this is only only for a backup, becase the csr
module is not suitable for forwarding data.

That's why test func_uty12_csrwr fails right now.

For example:

        addi.w   $r4, $r0, -1    # r4 = ffffffff
        csrwr    $r4, 0x0        # only crmd.ie (bit 2) can be set
        csrrd    $r5, 0x0        # crmd
        addi.w   $r5, $r5, 0x56

csrwr sets crmd to fffffff. Since there is only one bit register
that is crmd.ie, the crmd's value will be 00000004 and all other
fields are like read-only zeros.

Therefore, the next csrrd $r5, 0x0 will read 00000004 to r5.

But the data forwarding will forward ffffffff to r5.

This csr byp logic will be removed in the next commit.