add test case func_uty62_branch_next_instruction_executed_mistake
test if the instruction that follows a branch get executed by
mistake when the branch is taken.

When the IFU fetch instructon at each cycle, the pc_d shows the
instruction pc that follows the branch, but the inst_vld_d signal
is 0, because of kill_d.
So this instruction will not be executed, but this pc_d value will
pass through the pipeline.