add self defined csr register 0x100 bsec (Boot Security)
bit 0, ef eeprom flush

Once this bit is set to 1, it can not be set back to 0.

It is used in a scernario that the eeprom in the system can only
be flushed during the early boot stage. After the system
initialized, this bit is set and the eeprom is write protected.

add test case func_uty65_csr.bsec